requirements.txt
pandas
boto3
python-dotenv
sqlalchemy
psycopg2-binary
logging